+++ src/command_modules/azure-cli-iot/azure/cli/command_modules/iot/commands.py
@@ -14,18 +14,6 @@ class PolicyUpdateResultTransform(LongRunningOperation):  # pylint: disable=too-
         return result.properties.authorization_policies
 
 
-class LinkedHubUpdateResultTransform(LongRunningOperation):  # pylint: disable=too-few-public-methods
-    def __call__(self, poller):
-        result = super(LinkedHubUpdateResultTransform, self).__call__(poller)
-        return result.properties.iot_hubs
-
-
-class AccessPolicyUpdateResultTransform(LongRunningOperation):  # pylint: disable=too-few-public-methods
-    def __call__(self, poller):
-        result = super(AccessPolicyUpdateResultTransform, self).__call__(poller)
-        return result.properties.authorization_policies
-
-
 # Deleting IoT Hub is a long-running operation. Due to API implementation issue, 404 error will be thrown during
 # deletion of an IoT Hub.
 # This is a work around to suppress the 404 error. It should be removed after API is fixed.
@@ -57,20 +45,17 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
     with self.command_group('iot dps access-policy', client_factory=iot_service_provisioning_factory) as g:
         g.custom_command('list', 'iot_dps_access_policy_list')
         g.custom_command('show', 'iot_dps_access_policy_get')
-        g.custom_command('create', 'iot_dps_access_policy_create',
-                         transform=AccessPolicyUpdateResultTransform(self.cli_ctx))
-        g.custom_command('update', 'iot_dps_access_policy_update',
-                         transform=AccessPolicyUpdateResultTransform(self.cli_ctx))
-        g.custom_command('delete', 'iot_dps_access_policy_delete',
-                         transform=AccessPolicyUpdateResultTransform(self.cli_ctx))
+        g.custom_command('create', 'iot_dps_access_policy_create', no_wait_param='no_wait')
+        g.custom_command('update', 'iot_dps_access_policy_update', no_wait_param='no_wait')
+        g.custom_command('delete', 'iot_dps_access_policy_delete', no_wait_param='no_wait')
 
     # iot dps linked-hub commands
     with self.command_group('iot dps linked-hub', client_factory=iot_service_provisioning_factory) as g:
         g.custom_command('list', 'iot_dps_linked_hub_list')
         g.custom_command('show', 'iot_dps_linked_hub_get')
-        g.custom_command('create', 'iot_dps_linked_hub_create', transform=LinkedHubUpdateResultTransform(self.cli_ctx))
-        g.custom_command('update', 'iot_dps_linked_hub_update', transform=LinkedHubUpdateResultTransform(self.cli_ctx))
-        g.custom_command('delete', 'iot_dps_linked_hub_delete', transform=LinkedHubUpdateResultTransform(self.cli_ctx))
+        g.custom_command('create', 'iot_dps_linked_hub_create', no_wait_param='no_wait')
+        g.custom_command('update', 'iot_dps_linked_hub_update', no_wait_param='no_wait')
+        g.custom_command('delete', 'iot_dps_linked_hub_delete', no_wait_param='no_wait')
 
     # iot dps certificate commands
     with self.command_group('iot dps certificate', client_factory=iot_service_provisioning_factory) as g: